requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
Pillow==10.1.0
qrcode==7.4.2
python-barcode==0.15.1
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
twilio==8.10.3
flask-cors==4.0.0
bcrypt==4.1.2